adds :cache-key function option
The majority of the time the standard behaviour of using the args as the
cache key is fine.  However, for complex args (e.g. args which contain
 Java objects that break identity) you may need to specify how a unique
cache key is to be generated from the args.  (You may also want to do
this for storage/performance reasons.)  This commit gives you the
ability to specify a cach-key function that operates on the same args
which the value is computed from.