adds support for transient caches that can be GCed
Previous to this commit all caches were assumed to be resident in memory
for the entire lifetime of the process.  This makes since when you are
using def-cached or similar.  With this assumption it was safe to store
all the caches in the internal function-utils* atom.  This allowed for
easy invalidation of all defiened caches.

However, for temporary caches that are bound in a let or binding form you
do not want these caches to stick around after there lexical scopiing.
This commit allows the client to specify if a cache is intended to be
transient and thereby won't be added to the function-utils* atom.

This is an okay workaround, but not an ideal solution.  The ideal
solution would be to not force this extra call onto the client but to
always add the caches to function-utils* to keep the API consistent.
Instead of storing the objects, WeakReferences would need to be
used so that the caches would be GCed as normal.  I found this SO thread
about similar issues and Google' collections library was brought up as a
good option for this type of work:

http://stackoverflow.com/questions/3752957/thread-safe-object-pool-with-guarantee-of-returning-existing-object-if-strong-ref