fixes (lack of) serialization of key for ehcache, adds type hinting, and fixes a test
The type hinting was only added on the calls that are used repeatedly in
general caching operations.  These should be as fast as possible....

The test fixed was cached-names... It was fine until I added more caches
and then the test broke.  The test was relying on internal state of the
default CacheManager and would thereby be effected by all the other
tests.  So, I created a new manager jusr for that one test so it won't
break everytime a new cache is used in the test.  (A better approach probably
would be to reset the default cache manager for each test)